% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{class_performance}
\alias{class_performance}
\title{Class Performance data set}
\format{
A data frame with 3.618 rows and 6 columns
\describe{
   \item{student_id}{Unique ID number of the student - similar to the 
   staffer IDs generated by the `encode_staffer_ids` function,}
   \item{class_teacher}{Name of the faculty member teaching the class,}
   \item{quiz_number}{Quiz number N of the term,}
   \item{passed_quiz}{Boolean variable telling whether the student passed 
   (TRUE) or failed (FALSE) the quiz,}
   \item{passed_quiz_numeric}{0/1 whether the student passed or not,}
   \item{passed_quiz_string}{"Yes"/"No" string whether the student passed 
   the quiz or not.}
}
}
\usage{
class_performance
}
\description{
Example data set similar to one giving treatment outcomes in the SEPSIS
project. For each student/class_teacher combination, the data set gives
whether the student has passed Quiz #`test_number` by the second attempt.
This data set is used to test the `*_permutation_test*` functions
}
\keyword{datasets}
