% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_value.R
\name{get_p_value}
\alias{get_p_value}
\title{Permutation test helper}
\usage{
get_p_value(
  data_,
  ...,
  group_name = "group_",
  id_name = "id_",
  event_name = "event_",
  systematic = FALSE,
  na_fill = TRUE
)
}
\arguments{
\item{data_}{data frame with at least three columns: group, id and event}

\item{group_name}{name of the `group` column}

\item{id_name}{name of the `id` column}

\item{systematic}{whether the groups should be permuted consistently with 
the id, or not}
}
\value{
The test statistic after even groups have been permuted
}
\description{
standardises the input data table and retrieves the test statistic (p value)
}
\examples{
data("class_performance")

# calculating the p vakue from the permutation test with the included class_performance data set 
get_p_value(class_performance,
          group_name = group_name,
          id_name = id_name,
          event_name = event_name,
          systematic = TRUE,
          na_fill = TRUE)

# deprecated function; currently an alias for the get_p_value function 
permgp_fn(class_performance,
          group_name = group_name,
          id_name = id_name,
          event_name = event_name,
          systematic = TRUE,
          na_fill = TRUE)
}
