% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_permutation_test.R
\name{global_permutation_test}
\alias{global_permutation_test}
\title{Global Permutation Test}
\usage{
global_permutation_test(
  data_,
  group_col_name = "group_",
  id_col_name = "id_",
  event_col_name = "event_",
  ntrials = 10000,
  parallel = FALSE,
  ranseed = NaN,
  systematic = TRUE,
  na_fill = FALSE,
  check_ = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data_}{the data set, with at least three features: event (what 
happened), id (who were they), group (what intervention were they 
administered)}

\item{group_col_name}{the name of the column with `group` information}

\item{id_col_name}{the name of the column with participant/patient `id` 
information}

\item{event_col_name}{the name of the column with the event (data type 
castable to numerical)}

\item{ntrials}{the number of permutation tests performed (and test 
statistics calculated). currently capped at N!}

\item{parallel}{True/False whether the permutations/test statistics are to 
be calculated serially or in parallel (uses the R snow package)}

\item{ranseed}{a set random seed to be used for the permutations (used for 
reproducibility)}

\item{systematic}{True/False whether the groups are to be grouped or 
distributed totally randomly. For example, should all observations from 
Patient A in group (G1) be permuted to the same group (G2), or can they be 
spread across permuted groups? The first option (TRUE) is endorsed by 
Eleanor, the second option (FALSE) would be in line with other R packages 
(`coin`, for example).}

\item{na_fill}{if any entries in the input table have empty groups, either 
(TRUE) fill them with a generated group name .NA_group, or (FALSE) filter 
those rows out.}

\item{verbose}{TRUE/FALSE whether a completion message with stats should be
printed to the screen whenever the permutation test ends. A progress bar is
printed in all cases.}

\item{...}{currently ignored}
}
\value{
A list giving `$p` (the p-value), `$error` (the Monte-Carlo
error) of the calculation, `$N_trials` the number of permutations performed/
test statistics calculated, `$N_obs` the total number of observations for 
which the groups were permuted.
}
\description{
Carry through the permutation test
}
\examples{
# use the included class_performance data set for the example
data("class_performance")

# carry through the global permutation test on the class_performance data set
res <- global_permutation_test(class_performance,
                                group_col_name = "class_teacher",
                                id_col_name = "student_id",
                                event_col_name = "passed_quiz_numeric",
                                ntrials = 1000,
                                parallel = TRUE,
                                ranseed = NaN,
                                systematic = TRUE,
                                na_fill = TRUE,
                                verbose = TRUE)


# `res` is a list of four statistics
res
# $p
# [1] 0.337
# 
# $error
# [1] 0.01494761
# 
# $N_trials
# [1] 1000
# 
# $N_obs
# [1] 3618
}
