% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permut_tests.R
\name{perm_test_statistic}
\alias{perm_test_statistic}
\title{Test statistic calculation}
\usage{
perm_test_statistic(events, groups)
}
\arguments{
\item{events}{The scores to be parsed}

\item{groups}{The groups that the scores belong to}
}
\value{
The sum of square differences between group means and the overall
event mean
}
\description{
Calculate the test statistic (sum of square differences) between group means
and the overall mean
}
\examples{
# example for test statistic calculation
events <- c(1,0,0,1,1)
groups <- c(1,2,1,2,1)
perm_test_statistic(events, groups)
}
