% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permgp_fn.R
\name{permgp_fn}
\alias{permgp_fn}
\title{Permutation test helper (deprecated, please use `get_p_value`)}
\usage{
permgp_fn(
  data_,
  ...,
  group_col_name = "group_",
  id_col_name = "id_",
  event_col_name = "event_",
  systematic = FALSE,
  na_fill = TRUE
)
}
\arguments{
\item{data_}{data frame with at least three columns: group, id and event}

\item{group_col_name}{name of the `group` column}

\item{id_col_name}{name of the `id` column}

\item{systematic}{whether the groups should be permuted consistently with 
the id, or not}
}
\value{
The test statistic after even groups have been permuted
}
\description{
Standardises the input data table and retrieves the test statistic (p value)
}
\examples{
data("class_performance")

# calculating the p vakue from the permutation test with the included class_performance data set 
get_p_value(class_performance,
          group_col_name = group_col_name,
          id_col_name = id_col_name,
          event_col_name = event_col_name,
          systematic = TRUE,
          na_fill = TRUE)

# deprecated function; currently an alias for the get_p_value function 
permgp_fn(class_performance,
          group_col_name = group_col_name,
          id_col_name = id_col_name,
          event_col_name = event_col_name,
          systematic = TRUE,
          na_fill = TRUE)
}
