% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_groups.R
\name{permute_groups}
\alias{permute_groups}
\title{Group permutation}
\usage{
permute_groups(data_, systematic)
}
\arguments{
\item{data_}{data frame with at least three columns: group, id and event. this is created by the global_permutation_test function call}

\item{systematic}{whether the groups should be permuted consistently with 
the id, or not}
}
\value{
data frame (tibble) with both the original and permuted groups
}
\description{
function to carry out the group permutations
}
\examples{
"coming soon"
## standard col names, id_no, group_no, column_no

}
