% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SepsisTools-package.R
\docType{package}
\name{SepsisTools-package}
\alias{SepsisTools}
\alias{SepsisTools-package}
\title{SepsisTools: SepsisTools}
\description{
Code related to the SEPSIS project (2023-)
}
\author{
\strong{Maintainer}: Phillips Brendon \email{brendon.phillips@sickkids.ca}

Other contributors:
\itemize{
  \item Heasley Cole \email{cole.heasley@sickkids.ca} [contributor]
  \item Pullenayegum Eleanor \email{eleanor.pullenayegum@sickkids.ca} [contributor, translator]
  \item Roth Dan \email{daniel.roth@sickkids.ca} [contributor]
}

}
\keyword{internal}
