% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_permutation.R
\name{single_permutation}
\alias{single_permutation}
\title{Permutation test helper}
\usage{
single_permutation(
  data_,
  ...,
  group_col_name = "group_",
  id_col_name = "id_",
  event_col_name = "event_",
  systematic = FALSE,
  na_fill = TRUE,
  ranseed = NA
)
}
\arguments{
\item{data_}{the data set, with at least three features: event (what 
happened), id (who were they), group (what intervention were they 
administered)}

\item{...}{currently ignored}

\item{group_col_name}{the name of the column with `group` information}

\item{id_col_name}{the name of the column with participant/patient `id` 
information}

\item{event_col_name}{the name of the column with the event (data type 
castable to numerical)}

\item{systematic}{True/False whether the groups are to be grouped or 
distributed totally randomly. For example, should all observations from 
Patient A in group (G1) be permuted to the same group (G2), or can they be 
spread across permuted groups? The first option (TRUE) is endorsed by 
Eleanor, the second option (FALSE) would be in line with other R packages 
(`coin`, for example).}

\item{ranseed}{a set random seed to be used for the permutations (used for 
reproducibility)
#' @param na_fill if any entries in the input table have empty groups, either 
(TRUE) fill them with a generated group name .NA_group, or (FALSE) filter 
those rows out.}
}
\value{
The test statistic after even groups have been permuted
}
\description{
standardises the input data table and retrieves the test statistic (p value)
}
