% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SS_fn}
\alias{SS_fn}
\title{Test statistic calculation}
\usage{
SS_fn(event, group)
}
\arguments{
\item{event}{The scores to be parsed}

\item{group}{The groups that the scores belong to}
}
\value{
The temperature in degrees Celsius
}
\description{
Calculate the test statistic (sum of square differences)
}
\examples{
"coming soon"
}
