% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{staffer_dictionary}
\alias{staffer_dictionary}
\title{Staffer dictionary}
\format{
A data frame with 208 rows and 2 columns
\describe{
   \item{chrw_id}{The ID of the staffer appearing in the SEPSIS study data 
   set [not necessarily a community health research worker (CHRW)],}
   \item{encoding}{The new unique Roth Lab ID number that can be used to 
   track the worker through  out each encounter of the study.}
}
}
\usage{
staffer_dictionary
}
\description{
In the SEPSIS project, each staff member in every capacity (field, lab, etc) 
was given an ID number, usually of the form Dxxxxx/Nxxxxx (where each x is a 
digit 0-9). During QA, we found that these IDs were not unique over the span 
of the project and over all encounters, as was presumed. After communication 
with the investigative team in Bangladesh, we formulated a lookup table lut) 
that, based on our criteria, maps each data set ID number to a unique ID of 
the format `RLxxxxx` (where x is a digit 0-9, with 'RL' an abbreviation for 
"Roth Lab"). This lut is used in the function `encode_staffer_ids`.
}
\keyword{datasets}
