% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_table.R
\name{standard_table}
\alias{standard_table}
\title{Standardise the table}
\usage{
standard_table(
  data_,
  ...,
  id_col_name = "id_",
  group_col_name = "group_",
  event_col_name = "event_",
  id_vector = c(),
  group_vector = c(),
  event_vector = c(),
  na_fill = TRUE,
  data_frame_name = "input"
)
}
\arguments{
\item{data_}{lorem}

\item{...}{ipsum}

\item{group_vector}{si}

\item{event_vector}{amet}

\item{id_Vector}{dolor}
}
\value{
a standardised table to avoid faffing around with dynamic names
and such
}
\description{
Standardise the table
}
