% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permut_tests.R
\name{step_down_test}
\alias{step_down_test}
\title{Step-down test (Holm procedure)}
\usage{
step_down_test(first, second)
}
\arguments{
\item{first}{lorem}

\item{second}{ipsum}
}
\value{
p.values and etc.
}
\description{
Function to carry through the step-down procedure after the permutation test
}
\examples{
"coming soon"
 
}
