% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_chrw_ids.R
\name{uniquify_chrw}
\alias{uniquify_chrw}
\title{Uniquify Staffer IDs}
\usage{
uniquify_chrw(data_, ..., ID_col = NA, out_col_name = NA)
}
\arguments{
\item{data_}{data frame or list with the raw IDs to be uniquified}

\item{...}{currently ignored}

\item{ID_col}{the name of the column with the IDs to be uniquified (if `data_` is a table, ignored otherwise)
#' @param ID_col the name of the column with IDs to be encoded (ignored if .data is a list)}

\item{out_col_name}{the name of the column with the encoded IDs (if NA, the source column is overwritten and a column with a default name will store the original information)}
}
\value{
a list/data frame with the unique RothLab ID to allow tracing of activity throughout the encounters
}
\description{
A lookup junction to get unique IDs for each of the icddr,b staffers (field, lab, etc) that worked on the SEPSIS project and recorded information for any of the encounters
}
\examples{
"coming soon"

}
