% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniquify_staffer_ids.R
\name{uniquify_staffer_ids}
\alias{uniquify_staffer_ids}
\title{Uniquify Staffer IDs}
\usage{
uniquify_staffer_ids(data_, ..., ID_col = NA, out_col_name = NA)
}
\arguments{
\item{data_}{data frame or list with the raw IDs to be uniquified}

\item{...}{currently ignored}

\item{ID_col}{the name of the column with the IDs to be uniquified (if
`data_` is a table, ignored otherwise)
#' @param ID_col the name of the column with IDs to be encoded (ignored
if .data is a list)}

\item{out_col_name}{the name of the column with the encoded IDs (if NA,
the source column is overwritten and a column with a default name will
store the original information)}
}
\value{
a list/data frame with the unique RothLab ID to allow tracing of
activity throughout the encounters
}
\description{
A lookup function to get unique IDs for each of the icddr,b staffers
(field, lab, etc) that worked on the SEPSIS project and recorded
information for any of the encounters
}
\examples{
# uniquify a single staffer
uniquify_staffer_ids("D15248")

# uniquify a list of staffers
uniquify_staffer_ids(sample(staffer_dictionary$chrw_id, 20))

# uniquify a table, create a new column with the Roth Lab IDs
uniquify_staffer_ids(
  tibble(
    hehe = staffer_dictionary \%>\% pull(chrw_id) \%>\% sample(15),
    fake_col = sample(letters, 15, replace = TRUE)
  ),
  ID_col = hehe,
  out_col_name = hahalol
)

# the overwrite case - uniquify staffer IDs in a table, but ovewrite 
# the given ID column
uniquify_staffer_ids(
  tibble(
    lol = staffer_dictionary \%>\% pull(chrw_id) \%>\% sample(15),
    fake_col = sample(letters, 15, replace = TRUE)
  ),
  ID_col = lol
)
}
