% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_staffer_ids.R
\name{encode_staffer_ids}
\alias{encode_staffer_ids}
\title{String encoder}
\usage{
encode_staffer_ids(
  data_,
  ...,
  id_col = NA,
  out_col_name = NA,
  seed = NaN,
  na_fill = FALSE
)
}
\arguments{
\item{...}{currently ignored}

\item{id_col}{the name of the column with IDs to be encoded (ignored
if .data is a list)}

\item{out_col_name}{the name of the column with the encoded IDs (if NA,
the source column is overwritten and a column with a default name will
store the original information)}

\item{seed}{random seed to use (for reproducibility of results),}

\item{na_fill}{treat NA as a group and give them an ID}

\item{.data}{(optional) the vector/list/data frame with ids}
}
\value{
a data frame / list with id_col replaced with new Sepsis IDs, with
pre-encoding column appended to the end with Sys.time suffix
}
\description{
encodes any string into a standardised RTxxxxx format
}
\examples{
# load the `class_performance` test data set in the package
data(class_performance)

# encode a number
encode_staffer_ids(2)

# encode a list of strings
encode_staffer_ids(list("lisa", "daniel", "diego", "eleanor"))

# encode a vector of numbers with random seed given for reproducibility
encode_staffer_ids(1:10, seed = 10)

# encode a column of a table, replace the column (out_col_name not given)
encode_staffer_ids(class_performance, ID_col = "class_teacher", seed = 303902)

# encode a column of a table with given out_col name
encode_staffer_ids(class_performance,
  ID_col = "student_id",
  out_col_name = "encoding",
  seed = 303902
)

# using a magrittr pipe, replace column
class_performance \%>\% encode_staffer_ids(ID_col = "student_id", seed = 29)

# using a magrittr pipe, without a seed, encodings in aother column
class_performance \%>\% encode_staffer_ids(
  ID_col = "student_id",
  out_col_name = "encoding"
)
}
