% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_value.R
\name{get_p_value}
\alias{get_p_value}
\title{Permutation test helper}
\usage{
get_p_value(
  data_,
  ...,
  group_col_name = "group_",
  id_col_name = "id_",
  event_col_name = "event_",
  systematic = FALSE,
  na_fill = TRUE,
  ranseed = NA
)
}
\arguments{
\item{data_}{data frame with at least three columns: group, id and event}

\item{group_col_name}{name of the `group` column}

\item{id_col_name}{name of the `id` column}

\item{systematic}{whether the groups should be permuted consistently with
the id, or not}

\item{ranseed}{}
}
\value{
The test statistic after even groups have been permuted
}
\description{
standardises the input data table and retrieves the test statistic (p value)
}
