% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_permutation_test.R
\name{global_permutation_test}
\alias{global_permutation_test}
\title{Global Permutation Test}
\usage{
global_permutation_test(
  data_,
  ...,
  group_name = "group",
  id_name = "id",
  event_name = "event",
  ntrials = 10000,
  parallel = FALSE,
  ranseed = NaN,
  systematic = TRUE,
  na_fill = FALSE,
  check_ = FALSE
)
}
\arguments{
\item{data_}{data frame with at least 3 columns: event, id, group}

\item{group_name}{the name of the (IP ) column}

\item{id_name}{name of the (participant) id column}

\item{event_name}{name of the event column}

\item{ntrials}{the number of permutation tests in the ensemble}

\item{parallel}{True/False whether the permutations shosuld be run in
parallel for speed}

\item{ranseed}{fixed random seed, for reproducibility}

\item{systematic}{True/False whether the groups are to be grouped (similar
to the original entries) or totally random}

\item{na_fill}{if any entries in the input table have empty groups, fill them with a generated group name, or filter those rows out}
}
\value{
A list giving `$p` (the p-value) and `$error` (the Monte-Carlo error) of the calculation, `$N` the number of test statistics calculated
}
\description{
Carry through the permutation test
}
\examples{
"coming soon"

}
