% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_permutation_test.R
\name{global_permutation_test}
\alias{global_permutation_test}
\title{Global Permutation Test}
\usage{
global_permutation_test(
  data_,
  ...,
  group_col_name = "group_",
  id_col_name = "id_",
  event_col_name = "event_",
  ntrials = 10000,
  parallel = FALSE,
  ranseed = NaN,
  systematic = TRUE,
  na_fill = FALSE,
  check_ = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data_}{data frame with at least 3 columns: event (yes/no/numerical describing what happened), id (unique id of the patient/participant), group (what trial/observation group they belonged to)}

\item{group_col_name}{the name of the (IP) column}

\item{id_col_name}{name of the (participant) id column}

\item{event_col_name}{name of the event column}

\item{ntrials}{the number of permutation tests in the ensemble}

\item{parallel}{True/False whether the permutations should be run in
parallel for speed}

\item{ranseed}{possibility for using a fixed random seed, for reproducibility}

\item{systematic}{True/False whether the groups are to be grouped or distributed totally randomly. For example, should all observations from Patient A in group (G1) be permuted to the same group (G2), or can they be sporead across permuted groups? The second would be in line with other prepackaged R libraries.}

\item{na_fill}{if any entries in the input table have empty groups, fill
them with a generated group name (.NA_group), or filter those rows out}

\item{verbose}{TRUE/FALSE whether a completion message with stats should be
printed to the screen whenever the permutation test ends. A progress bar is
printed in all cases.}
}
\value{
A list giving `$p` (the p-value) and `$error` (the Monte-Carlo
error) of the calculation, `$N_perms` the number of permutations performed/test statistics calculated, `$N_trials` the total number of observations for which the groups were permuted.
}
\description{
Carry through the permutation test
}
\examples{
"coming soon"

}
