% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permut_tests.R
\name{pairwise_permutation_tests}
\alias{pairwise_permutation_tests}
\title{Pairwise permutation tests}
\usage{
pairwise_permutation_tests(first, second, third)
}
\arguments{
\item{first}{lorem}

\item{second}{ipsum}
}
\value{
returning p values from pairwise tests
}
\description{
Conduct all possible permutation tests
}
\examples{
"coming soon"

}
