% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_permutation_test.R
\name{pairwise_permutation_tests}
\alias{pairwise_permutation_tests}
\title{Pairwise permutation tests}
\usage{
pairwise_permutation_tests(
  data_,
  ...,
  group_col_name = "group",
  id_col_name = "id",
  event_col_name = "event",
  ntrials = 100,
  parallel = FALSE,
  ranseed = NaN,
  systematic = TRUE,
  na_fill = FALSE,
  backsolve = FALSE,
  reference_group = "placebo",
  step_down_procedure = "HB",
  global_test_first = TRUE
)
}
\arguments{
\item{first}{lorem}

\item{second}{ipsum}
}
\value{
returning p values from pairwise tests
}
\description{
Conduct all possible permutation tests
}
\examples{
"coming soon"

}
