% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permut_tests.R
\name{permgp_fn}
\alias{permgp_fn}
\title{Permutation test helper}
\usage{
permgp_fn(data_, group_name, id_name, event_name, systematic)
}
\arguments{
\item{data_}{data frame with at least three columns: group, id and event}

\item{group_name}{name of the `group` column}

\item{id_name}{name of the `id` column}

\item{systematic}{whether the groups should be permuted consistently with 
the id, or not}
}
\value{
The test statistic after even groups have been permuted
}
\description{
helper function for permutation test
}
\examples{
"coming soon"
}
