% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permgp_fn.R
\name{permgp_fn}
\alias{permgp_fn}
\title{Permutation test helper (deprecated, please use `single_permutation`)}
\usage{
permgp_fn(
  data_,
  ...,
  group_col_name = "group_",
  id_col_name = "id_",
  event_col_name = "event_",
  systematic = FALSE,
  na_fill = TRUE,
  ranseed = NA
)
}
\description{
Permutation test helper (deprecated, please use `single_permutation`)
}
\examples{
data("class_performance")

# calculating the p vakue from the permutation test with the included class_performance data set 
single_permutation(class_performance,
          group_col_name = group_col_name,
          id_col_name = id_col_name,
          event_col_name = event_col_name,
          systematic = TRUE,
          na_fill = TRUE)

# deprecated function; currently an alias for the single_permutation function 
permgp_fn(class_performance,
          group_col_name = group_col_name,
          id_col_name = id_col_name,
          event_col_name = event_col_name,
          systematic = TRUE,
          na_fill = TRUE)
}
