% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SepsisTools-package.R
\docType{package}
\name{SepsisTools-package}
\alias{SepsisTools}
\alias{SepsisTools-package}
\title{SepsisTools: SepsisTools}
\description{
Code related to the SEPSIS project (2023-)
}
\author{
\strong{Maintainer}: Brendon Phillips \email{brendon.phillips@sickkids.ca}

Other contributors:
\itemize{
  \item Cole Heasley \email{cole.heasley@sickkids.ca} [contributor]
  \item Eleanor Pullenayegum \email{eleanor.pullenayegum@sickkids.ca} [contributor, translator]
  \item Dan Roth \email{daniel.roth@sickkids.ca} [contributor]
}

}
\keyword{internal}
